Update to scene and object properties/behaviors
Cleaned up building of scenes and objects with properties, controllers, models, and views. Descriptions were added for specific functionalities, more properties were defined, and some properties were renamed for clarity, much of the positioning and manipulations methods underwent refactoring for more efficient behavior. Removal of the camera from the testing of scenes was necessary as it is no longer a suitable test method. Overall, this change set should improve the clarity and functionality, and provide a more solid base for future development.