Add MIDI song creation and playing capabilities
Implement audio component in the game engine, focusing specifically on MIDI song playing. This includes the MIDI song factory, song player interface as well as their implementations as classes in the `Audio` project. Generated song definitions using source generators to automatically include the MIDI files from the Assets folder into the build. This allows the song to be changed / added simply by changing / adding the corresponding MIDI file. The solution file and related project files are also updated to include and manage the newly added classes and files. Finally, provided a console audio player with simple controls and tune library to demonstrate the audio playback capabilities in a console environment. This improvement is critical for audio support in the game engine, enhancing the game atmosphere and player experience.